backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
PyJWT==2.8.0
requests==2.31.0
